src/backend/containers/query-gedi/requirements.txt
# Core dependencies for Python 3.8
numpy==1.24.4
scipy==1.10.1
open3d==0.15.2
scikit-learn==1.3.2

# GeDi specific dependencies
torchgeometry==0.1.2

# System libraries
psutil==5.9.6
tqdm==4.66.1
pillow==10.4.0
matplotlib==3.7.5
PyYAML==6.0.2

# Kafka integration
kafka-python==2.0.2
flask==2.3.2

# Performance libraries
joblib==1.4.2
GPUtil==1.4.0
open3d==0.15.2
torchgeometry==0.1.2
gdown
tensorboard
protobuf==3.20
